requirements.txt
uwsgi
psycopg2-binary
django==4.2.4
django-compressor
requests
djangorestframework
beautifulsoup4
tzdata
django-cors-headers
django_apscheduler
prettytable
django-crontab
python-telegram-bot
